rename results-view folder, move views
