+++ apps/roam/src/components/DiscourseContext.tsx
@@ -20,7 +20,7 @@ import getShallowTreeByParentUid from "roamjs-components/queries/getShallowTreeB
 import { Result } from "roamjs-components/types/query-builder";
 import nanoId from "nanoid";
 import getDiscourseContextResults from "../utils/getDiscourseContextResults";
-import ResultsView from "./ResultsView/ResultsView";
+import ResultsView from "./results-view/ResultsView";
 import { getPageTitleValueByHtmlElement } from "roamjs-components/dom";
 import getPageUidByPageTitle from "roamjs-components/queries/getPageUidByPageTitle";
 import renderWithUnmount from "roamjs-components/util/renderWithUnmount";
