+++ apps/roam/src/components/results-view/ResultsView.tsx
@@ -27,16 +27,16 @@ import { useExtensionAPI } from "roamjs-components/components/ExtensionApiContex
 import postProcessResults from "~/utils/postProcessResults";
 import setInputSetting from "roamjs-components/util/setInputSetting";
 import getUids from "roamjs-components/dom/getUids";
-import Charts from "./views/Charts";
-import Timeline from "./views/Timeline";
-import Kanban from "./views/Kanban";
+import Charts from "./Charts";
+import Timeline from "./Timeline";
+import Kanban from "./Kanban";
 import MenuItemSelect from "roamjs-components/components/MenuItemSelect";
 import { RoamBasicNode } from "roamjs-components/types";
 import { render as renderToast } from "roamjs-components/components/Toast";
 import { Column, Result } from "~/utils/types";
 import updateBlock from "roamjs-components/writes/updateBlock";
 import { Condition } from "~/utils/types";
-import ResultsTable from "./views/ResultsTable";
+import ResultsTable from "./ResultsTable";
 import { render as renderSimpleAlert } from "roamjs-components/components/SimpleAlert";
 import setInputSettings from "roamjs-components/util/setInputSettings";
 import posthog from "posthog-js";
